текущая линейка умеет 75% границ. добавлены тесты

+++ FiveDiff/Program.cs
@@ -9,7 +9,9 @@ namespace FiveDiff
 {
     static class Program
     {
-
+        // заполнится ниже
+        static public string TempFolder = "";
+        // элементы формы
         static public PictureBox img;
         static public ComboBox SelectLang;
         static public ComboBox SelectNumPlace;
@@ -17,8 +19,16 @@ namespace FiveDiff
         static public TextBox Answer;
         static public Form F;
         static public System.Timers.Timer t1;
+        // ответ картинками для отображения
         static public Bitmap[] pairs;
+        // флаг для таймера
         static public int num = 1;
+
+        //
+        //
+        // удалить после перехода на новый класс
+        //
+        //
         static public string answerformat;// = "LaUp1A";
 
         /// <summary>
@@ -30,17 +40,70 @@ namespace FiveDiff
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
 
+            TempFolder = System.IO.Path.GetTempPath();
             PrepareForm();
             PrepareTimer();
 
-            string v1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\kmv57389_EnUpA1\A5B2B4B7E6-2.jpg";
-            string v2 = "EnUpA1";
-            FindFiveDiff a = new FindFiveDiff(new Bitmap(v1), v2);
+            string p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\pic02.jpg"; // зеленая игра
+            //p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\9\lu.jpg";
+            p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\5_EnUpA1\B5C1F6H8L4.jpg"; //розовая сетка
+            //p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\5_EnUpA1\A2A7D2D5E8.jpg"; // яйца, поля
+            //p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\4_EnLeA1\F2H3C6G8D9.jpg"; // шахматы
+            p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\sar21623_RuLeA1\Б3Д4А6И6Ж8-2.jpg";
+            p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\B6C3C4D2E5-1.jpg"; // фрукты
+            p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\A2A7D2D5E8.jpg"; // белая раска вокруг
+            p1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\A4F4I5J6K3.jpg"; // белый фон картинки
+            FiveDifference b = new FiveDifference(new Bitmap(p1), new FiveDifference.Settings(FiveDifference.setNums.Left, FiveDifference.setLang.Rus, FiveDifference.setType._A1));
 
-            F_SizeChanged(null, null);
-            Application.Run(F);
+            //string v1 = @"C:\Users\Антон\Source\Repos\FiveDiff\FiveDiff\FiveDiffTests\pics4test\kmv57389_EnUpA1\A5B2B4B7E6-2.jpg";
+            //string v2 = "EnUpA1";
+            //FindFiveDiff a = new FindFiveDiff(new Bitmap(v1), v2);
+
+            //F_SizeChanged(null, null);
+            //Application.Run(F);
         }
 
+        //
+        //
+        //
+        // здесь и в следующем методе необходимо будет заменить метод подготовки формата ответа и вызова класса, решающего задачу.
+        //
+        //
+        //
+        private static void CreateAnswerFormat()
+        {
+            string lang = "La";
+            if (SelectLang.SelectedIndex == 1) { lang = "Ru"; }
+            string nums = "Up";
+            if (SelectNumPlace.SelectedIndex == 1) { nums = "Le"; }
+            string answ = "1A";
+            if (SelectAnswerType.SelectedIndex == 1) { answ = "A1"; }
+            answerformat = lang + nums + answ;
+        }
+        private static void RunWithFile(string v)
+        {
+            using (Bitmap Bmp = new Bitmap(v))
+            {
+                img.Image = Bmp;
+                F.Update();
+                CreateAnswerFormat();
+                FindFiveDiff a = new FindFiveDiff(Bmp, answerformat);
+                pairs = new Bitmap[2];
+                pairs[0] = a.Img1;
+                pairs[1] = a.Img2;
+                Answer.Text = a.Answer;
+                Clipboard.SetText(a.Answer);
+                a = null;
+            }
+        }
+        //
+        //
+        //
+
+
+        /// <summary>
+        /// подготовка таймера
+        /// </summary>
         private static void PrepareTimer()
         {
             t1 = new System.Timers.Timer();
@@ -48,7 +111,9 @@ namespace FiveDiff
             t1.AutoReset = true;
             t1.Elapsed += T_Elapsed;
         }
-
+        /// <summary>
+        /// подготовка формы
+        /// </summary>
         private static void PrepareForm()
         {
             F = new Form();
@@ -68,8 +133,8 @@ namespace FiveDiff
             //SelectNumPlace.SelectedIndexChanged += Select_SelectedIndexChanged;
             F.Controls.Add(SelectNumPlace);
             SelectAnswerType = new ComboBox();
-            SelectAnswerType.Items.Add("1A2B3C4D5E");
             SelectAnswerType.Items.Add("A1B2C3D4E5");
+            SelectAnswerType.Items.Add("1A2B3C4D5E");
             SelectAnswerType.SelectedIndex = 0;
             //SelectAnswerType.SelectedIndexChanged += Select_SelectedIndexChanged;
             F.Controls.Add(SelectAnswerType);
@@ -102,29 +167,30 @@ namespace FiveDiff
             img.DragEnter += Img_DragEnter;
             F.Controls.Add(img);
         }
-
-        private static void CreateAnswerFormat()
-        {
-            string lang = "La";
-            if (SelectLang.SelectedIndex == 1) { lang = "Ru"; }
-            string nums = "Up";
-            if (SelectNumPlace.SelectedIndex == 1) { nums = "Le"; }
-            string answ = "1A";
-            if (SelectAnswerType.SelectedIndex == 1) { answ = "A1"; }
-            answerformat = lang + nums + answ;
-        }
-
+        /// <summary>
+        /// дейсвтие таймера (менять картинку)
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private static void T_Elapsed(object sender, System.Timers.ElapsedEventArgs e)
         {
             img.Image = pairs[num];
             num = 1 - num;
         }
-
+        /// <summary>
+        /// d&d enter event
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private static void Img_DragEnter(object sender, DragEventArgs e)
         {
             e.Effect = DragDropEffects.Move;
         }
-
+        /// <summary>
+        /// ивент на перетаскивание файла или ссылки
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private static void Img_DragDrop(object sender, DragEventArgs e)
         {
             t1.Enabled = false;
@@ -132,10 +198,10 @@ namespace FiveDiff
             if (lnk != null)
             {
                 // перетащенная ссылка из браузера
-                using (var client = new System.Net.WebClient())
+                string path = TempFolder + lnk.Substring(lnk.LastIndexOf("/") + 1);
+                SafeDownload(lnk, path);
+                if (System.IO.File.Exists(path))
                 {
-                    string path = @"C:\TEMP\" + lnk.Substring(lnk.LastIndexOf("/") + 1);
-                    client.DownloadFile(lnk, path);
                     RunWithFile(path);
                     System.IO.File.Delete(path);
                 }
@@ -148,24 +214,36 @@ namespace FiveDiff
             }
             t1.Enabled = true;
         }
-
-        private static void RunWithFile(string v)
+        /// <summary>
+        /// пятикратные попытки скачивания - не гарант, но хоть чтото
+        /// </summary>
+        /// <param name="lnk">ссылка</param>
+        /// <param name="path">путь для сохранения локально</param>
+        private static void SafeDownload(string lnk, string path)
         {
-            using (Bitmap Bmp = new Bitmap(v))
+            using (var client = new System.Net.WebClient())
             {
-                img.Image = Bmp;
-                F.Update();
-                CreateAnswerFormat();
-                FindFiveDiff a = new FindFiveDiff(Bmp, answerformat);
-                pairs = new Bitmap[2];
-                pairs[0] = a.Img1;
-                pairs[1] = a.Img2;
-                Answer.Text = a.Answer;
-                Clipboard.SetText(a.Answer);
-                a = null;
+                int max_count = 5;
+                bool need_download = true;
+                while (need_download && (max_count != 0))
+                {
+                    try
+                    {
+                        client.DownloadFile(lnk, path);
+                        need_download = false;
+                    }
+                    catch
+                    {
+                        max_count--;
+                    }
+                }
             }
         }
-
+        /// <summary>
+        /// изменение размеров окна-  корректировка положения картинки
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private static void F_SizeChanged(object sender, EventArgs e)
         {
             img.Top = 5;
